+++ services/secretsmanager/src/stackit/secretsmanager/api/default_api.py
@@ -113,6 +113,7 @@ class DefaultApi:
             "400": "BadRequest",
             "401": "str",
             "404": "NotFound",
+            "409": "Conflict",
             "500": None,
         }
         response_data = self.api_client.call_api(*_param, _request_timeout=_request_timeout)
@@ -187,6 +188,7 @@ class DefaultApi:
             "400": "BadRequest",
             "401": "str",
             "404": "NotFound",
+            "409": "Conflict",
             "500": None,
         }
         response_data = self.api_client.call_api(*_param, _request_timeout=_request_timeout)
@@ -261,6 +263,7 @@ class DefaultApi:
             "400": "BadRequest",
             "401": "str",
             "404": "NotFound",
+            "409": "Conflict",
             "500": None,
         }
         response_data = self.api_client.call_api(*_param, _request_timeout=_request_timeout)
@@ -3626,6 +3629,7 @@ class DefaultApi:
             "400": "BadRequest",
             "401": "str",
             "404": "NotFound",
+            "409": "Conflict",
             "500": None,
         }
         response_data = self.api_client.call_api(*_param, _request_timeout=_request_timeout)
@@ -3700,6 +3704,7 @@ class DefaultApi:
             "400": "BadRequest",
             "401": "str",
             "404": "NotFound",
+            "409": "Conflict",
             "500": None,
         }
         response_data = self.api_client.call_api(*_param, _request_timeout=_request_timeout)
@@ -3774,6 +3779,7 @@ class DefaultApi:
             "400": "BadRequest",
             "401": "str",
             "404": "NotFound",
+            "409": "Conflict",
             "500": None,
         }
         response_data = self.api_client.call_api(*_param, _request_timeout=_request_timeout)