+++ services/secretsmanager/src/stackit/secretsmanager/models/kms_key_payload.py
@@ -17,20 +17,23 @@ import json
 import pprint
 from typing import Any, ClassVar, Dict, List, Optional, Set
 
-from pydantic import BaseModel, ConfigDict, Field, StrictStr
+from pydantic import BaseModel, ConfigDict, Field, StrictInt, StrictStr
 from typing_extensions import Self
 
-from stackit.secretsmanager.models.kms_key_payload import KmsKeyPayload
 
-
-class UpdateInstancePayload(BaseModel):
+class KmsKeyPayload(BaseModel):
     """
-    UpdateInstancePayload
+    The key for secret encryption and decryption.
     """  # noqa: E501
 
-    kms_key: Optional[KmsKeyPayload] = Field(default=None, alias="kmsKey")
-    name: StrictStr = Field(description="A user chosen name to distinguish multiple secrets manager instances.")
-    __properties: ClassVar[List[str]] = ["kmsKey", "name"]
+    key_id: StrictStr = Field(description="The key UUID.", alias="keyId")
+    key_ring_id: StrictStr = Field(description="The key ring UUID the key is part of.", alias="keyRingId")
+    key_version: StrictInt = Field(description="The Key version number.", alias="keyVersion")
+    service_account_email: StrictStr = Field(
+        description="The Service account email that will consume the key. Must be in the same project as the Secrets Manager instance.",
+        alias="serviceAccountEmail",
+    )
+    __properties: ClassVar[List[str]] = ["keyId", "keyRingId", "keyVersion", "serviceAccountEmail"]
 
     model_config = ConfigDict(
         populate_by_name=True,
@@ -49,7 +52,7 @@ class UpdateInstancePayload(BaseModel):
 
     @classmethod
     def from_json(cls, json_str: str) -> Optional[Self]:
-        """Create an instance of UpdateInstancePayload from a JSON string"""
+        """Create an instance of KmsKeyPayload from a JSON string"""
         return cls.from_dict(json.loads(json_str))
 
     def to_dict(self) -> Dict[str, Any]:
@@ -69,14 +72,11 @@ class UpdateInstancePayload(BaseModel):
             exclude=excluded_fields,
             exclude_none=True,
         )
-        # override the default output from pydantic by calling `to_dict()` of kms_key
-        if self.kms_key:
-            _dict["kmsKey"] = self.kms_key.to_dict()
         return _dict
 
     @classmethod
     def from_dict(cls, obj: Optional[Dict[str, Any]]) -> Optional[Self]:
-        """Create an instance of UpdateInstancePayload from a dict"""
+        """Create an instance of KmsKeyPayload from a dict"""
         if obj is None:
             return None
 
@@ -85,8 +85,10 @@ class UpdateInstancePayload(BaseModel):
 
         _obj = cls.model_validate(
             {
-                "kmsKey": KmsKeyPayload.from_dict(obj["kmsKey"]) if obj.get("kmsKey") is not None else None,
-                "name": obj.get("name"),
+                "keyId": obj.get("keyId"),
+                "keyRingId": obj.get("keyRingId"),
+                "keyVersion": obj.get("keyVersion"),
+                "serviceAccountEmail": obj.get("serviceAccountEmail"),
             }
         )
         return _obj